requirements.txt
PyQt6>=6.6.0
mss>=9.0.0
Pillow>=10.0.0
pynput>=1.7.6
pywin32>=306
